Adding rotor status functionality to web UI
added code to query the rotor AZ/EL (via hamlib and the configured
ROT_IP and ROT_PORT). If successful then the status turns green
and the AZ/EL of the ground station is updated. If anything fails
along the way (port unavailable, reponse is not the float we
expect, etc) then the AZ/EL is reset to NA and the status turns
red.